requirements.txt
aksharamukha==2.2.2
annotated-types==0.7.0
anyio==4.6.0
backports.functools-lru-cache==2.0.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.11
Deprecated==1.2.14
fastapi==0.112.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
h11==0.14.0
idna==3.10
indic-transliteration==2.3.61
inexactsearch==1.0.2
jaconv==0.4.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
openai-whisper @ git+https://github.com/openai/whisper.git@279133e3107392276dc509148da1f41bfb532c7e
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pykakasi==2.3.0
python-multipart==0.0.10
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
roman==4.2
scipy==1.14.1
shellingham==1.5.4
silpa_common==0.3
six==1.16.0
sniffio==1.3.1
soundex==1.1.3
spellchecker==0.4
starlette==0.37.2
sympy==1.13.3
textblob==0.18.0.post0
tiktoken==0.7.0
toml==0.10.2
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
unicodedata2==15.1.0
urllib3==2.2.3
uvicorn==0.30.6
websockets==13.1
wrapt==1.16.0
